% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spgl.dt.R
\name{spgl.dt.dtDlDir}
\alias{spgl.dt.dtDlDir}
\title{spgl.dt.dtDlDir}
\usage{
spgl.dt.dtDlDir(id = integer(), forDate = as_date(character()),
  dlDir = character(), dlDir.fllPath = character())
}
\arguments{
\item{id}{integer}

\item{forDate}{as_date(character())}

\item{dlDir}{character}

\item{dlDir.fllPath}{character}
}
\value{
data.table
}
\description{
This function creates an object of spgl.dt.dtDlDir.
}
\details{
XXX
}
\examples{
dtDlDir1 <- spgl.dt.dtDlDir()
print(dtDlDir1)

dtDlDir2 <- spgl.dt.dtDlDir(id = as.integer(1),
                           forDate = as_date(today()),
                           dlDir = "a",
                           dlDir.fllPath = "b")
print(dtDlDir2)
}
\author{
Christian Frei
}
