% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spgl.dt.R
\name{spgl.dt.dtHeadline}
\alias{spgl.dt.dtHeadline}
\title{spgl.dt.dtHeadline}
\usage{
spgl.dt.dtHeadline(id = integer(), idHeadline = integer(),
  idHeadlineFile = integer(), forDate = lubridate::as_date(character()),
  forTime = lubridate::as_datetime(character()), isBentoLink = logical(),
  isSpiegelPlusLink = logical(), intro = character(), title = character(),
  section = character(), link = character())
}
\arguments{
\item{id}{integer}

\item{idHeadline}{integer}

\item{idHeadlineFile}{integer}

\item{forDate}{POSIXt or character}

\item{forTime}{POSIXt or character}

\item{isBentoLink}{logical}

\item{isSpiegelPlusLink}{logical}

\item{intro}{character}

\item{title}{character}

\item{section}{character}

\item{link}{character}
}
\value{
data.table
}
\description{
This function creates an object of spgl.dt.dtHeadline.
}
\examples{
dtHeadline1 <- spgl.dt.dtHeadline()
print(dtHeadline1)

dtHeadline2 <- spgl.dt.dtHeadline(id = as.integer(1),
                                  idHeadline = as.integer(2),
                                  idHeadlineFile = as.integer(3),
                                  forDate = as_date(today()),
                                  forTime = as_datetime(now()),
                                  isBentoLink = FALSE,
                                  isSpiegelPlusLink = FALSE,
                                  intro = "a",
                                  title = "b",
                                  section = "c",
                                  link = "d")
print(dtHeadline2)
}
\author{
Christian Frei
}
