% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spgl.dt.R
\name{spgl.dt.dtHeadlineFile}
\alias{spgl.dt.dtHeadlineFile}
\title{spgl.dt.dtHeadlineFile}
\usage{
spgl.dt.dtHeadlineFile(id = integer(), idDlDir = integer(),
  forDate = lubridate::as_date(character()),
  inDate = lubridate::as_date(character()),
  inTime = lubridate::as_datetime(character()), dlDir = character(),
  dlDir.fllPath = character(), rawDir = character(),
  rawDir.fllPath = character(), fileHTML = character(),
  fileHTML.fllPath = character(), fileCSV = character(),
  fileCSV.fllPath = character())
}
\arguments{
\item{id}{integer}

\item{idDlDir}{integer}

\item{forDate}{POSIXt or character}

\item{inDate}{POSIXt or character}

\item{inTime}{POSIXt or character}

\item{dlDir}{character}

\item{dlDir.fllPath}{character}

\item{rawDir}{character}

\item{rawDir.fllPath}{character}

\item{fileHTML}{character}

\item{fileHTML.fllPath}{character}

\item{fileCSV}{character}

\item{fileCSV.fllPath}{character}
}
\value{
data.table
}
\description{
This function creates an object of spgl.dt.dtHeadlineFile.
}
\details{
XXX
}
\examples{
dtHeadlineFile1 <- spgl.dt.dtHeadlineFile()
print(dtHeadlineFile1)

dtHeadlineFile2 <- spgl.dt.dtHeadlineFile(id                   = as.integer(1),
                                          idDlDir              = as.integer(1),
                                          forDate              = lubridate::as_date(today()),
                                          inDate               = lubridate::as_date(today()),
                                          inTime               = lubridate::as_datetime(now()),
                                          dlDir                = "a",
                                          dlDir.fllPath        = "b",
                                          rawDir               = "c",
                                          rawDir.fllPath       = "d",
                                          fileHTML             = "e",
                                          fileHTML.fllPath     = "f",
                                          fileCSV              = "g",
                                          fileCSV.fllPath      = "h")
print(dtHeadlineFile2)
}
\author{
Christian Frei
}
