% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spgl.dt.R
\docType{methods}
\name{spgl.dt.lsFileStr.addDtDlDir}
\alias{spgl.dt.lsFileStr.addDtDlDir}
\alias{spgl.dt.lsFileStr.addDtDlDir,spgl.dt.lsFileStr,data.table-method}
\title{spgl.dt.lsFileStr.addDtDlDir}
\usage{
spgl.dt.lsFileStr.addDtDlDir(lsFileStr, dtDlDir)


  \S4method{spgl.dt.lsFileStr.addDtDlDir}{spgl.dt.lsFileStr,data.table}(lsFileStr,
  dtDlDir)
}
\arguments{
\item{lsFileStr}{spgl.dt.lsFileStr.setDtDlDir}

\item{dtDlDir}{data.table}
}
\value{
spgl.dt.lsFileStr spgl.dt.lsFileStr
}
\description{
create a method to get the value of the dtDlDir
}
\details{
XXX
}
\examples{
lsFileStr <- new("spgl.dt.lsFileStr")
dtDlDir1 <- spgl.dt.dtDlDir(id = as.integer(1),
                           forDate = as_date(today()),
                           dlDir = "a",
                           dlDir.fllPath = "b")
lsFileStr <- spgl.dt.lsFileStr.setDtDlDir(lsFileStr, dtDlDir1)
dtDlDir2 <- spgl.dt.dtDlDir(id = as.integer(2),
                           forDate = as_date(today())-years(1),
                           dlDir = "y",
                           dlDir.fllPath = "z")
lsFileStr <- spgl.dt.lsFileStr.addDtDlDir(lsFileStr, dtDlDir2)
print(spgl.dt.lsFileStr.getDtDlDir(lsFileStr))
}
\author{
Christian Frei
}
