% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spgl.dt.R
\docType{methods}
\name{spgl.dt.lsFileStr.addDtHeadline}
\alias{spgl.dt.lsFileStr.addDtHeadline}
\alias{spgl.dt.lsFileStr.addDtHeadline,spgl.dt.lsFileStr,data.table-method}
\title{spgl.dt.lsFileStr.addDtHeadline}
\usage{
spgl.dt.lsFileStr.addDtHeadline(lsFileStr, dtHeadline)


  \S4method{spgl.dt.lsFileStr.addDtHeadline}{spgl.dt.lsFileStr,data.table}(lsFileStr,
  dtHeadline)
}
\arguments{
\item{lsFileStr}{spgl.dt.lsFileStr}

\item{dtHeadline}{dtHeadline}
}
\value{
spgl.dt.lsFileStr
}
\description{
xxx
}
\details{
XXX
}
\examples{
lsFileStr <- new("spgl.dt.lsFileStr")
dtHeadline1 <- spgl.dt.dtHeadline(id = as.integer(1),
                                  idHeadline = as.integer(2),
                                  idHeadlineFile = as.integer(3),
                                  forDate = as_date(today()),
                                  forTime = as_datetime(now()),
                                  isBentoLink = FALSE,
                                  isSpiegelPlusLink = FALSE,
                                  intro = "a",
                                  title = "b",
                                  section = "c",
                                  link = "d")
lsFileStr <- spgl.dt.lsFileStr.setDtHeadline(lsFileStr, dtHeadline1)
dtHeadline2 <- spgl.dt.dtHeadline(id = as.integer(4),
                                  idHeadline = as.integer(5),
                                  idHeadlineFile = as.integer(6),
                                  forDate = as_date(today()-years(1)),
                                  forTime = as_datetime(now()-years(1)),
                                  isBentoLink = FALSE,
                                  isSpiegelPlusLink = FALSE,
                                  intro = "e",
                                  title = "f",
                                  section = "g",
                                  link = "h")
lsFileStr <- spgl.dt.lsFileStr.addDtHeadline(lsFileStr, dtHeadline2)
print(spgl.dt.lsFileStr.getDtHeadline(lsFileStr))
}
\author{
Christian Frei
}
