% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spgl.dt.R
\docType{methods}
\name{spgl.dt.lsFileStr.addDtHeadlineFile}
\alias{spgl.dt.lsFileStr.addDtHeadlineFile}
\alias{spgl.dt.lsFileStr.addDtHeadlineFile,spgl.dt.lsFileStr,data.table-method}
\title{spgl.dt.lsFileStr.addDtHeadlineFile}
\usage{
spgl.dt.lsFileStr.addDtHeadlineFile(lsFileStr, dtHeadlineFile)


  \S4method{spgl.dt.lsFileStr.addDtHeadlineFile}{spgl.dt.lsFileStr,data.table}(lsFileStr,
  dtHeadlineFile)
}
\arguments{
\item{lsFileStr}{spgl.dt.lsFileStr.setDtDlDir}

\item{dtHeadlineFile}{data.table}
}
\value{
spgl.dt.lsFileStr spgl.dt.lsFileStr
}
\description{
create a method to get the value of the dtDlDir
}
\details{
XXX
}
\examples{
lsFileStr <- new("spgl.dt.lsFileStr")
dtHeadlineFile1 <- spgl.dt.dtHeadlineFile(id                   = as.integer(1),
                                          idDlDir              = as.integer(1),
                                          forDate              = lubridate::as_date(today()),
                                          inDate               = lubridate::as_date(today()),
                                          inTime               = lubridate::as_datetime(now()),
                                          dlDir                = "a",
                                          dlDir.fllPath        = "b",
                                          rawDir               = "c",
                                          rawDir.fllPath       = "d",
                                          fileHTML             = "e",
                                          fileHTML.fllPath     = "f",
                                          fileCSV              = "g",
                                          fileCSV.fllPath      = "h")
lsFileStr <- spgl.dt.lsFileStr.setDtHeadlineFile(lsFileStr, dtHeadlineFile1)
dtHeadlineFile2 <- spgl.dt.dtHeadlineFile(id                   = as.integer(2),
                                          idDlDir              = as.integer(2),
                                          forDate              = lubridate::as_date(today()-years(1)),
                                          inDate               = lubridate::as_date(today()-years(1)),
                                          inTime               = lubridate::as_datetime(now()-years(1)),
                                          dlDir                = "g",
                                          dlDir.fllPath        = "h",
                                          rawDir               = "i",
                                          rawDir.fllPath       = "j",
                                          fileHTML             = "k",
                                          fileHTML.fllPath     = "l",
                                          fileCSV              = "m",
                                          fileCSV.fllPath      = "n")
lsFileStr <- spgl.dt.lsFileStr.addDtHeadlineFile(lsFileStr, dtHeadlineFile2)
print(spgl.dt.lsFileStr.getDtHeadlineFile(lsFileStr))
}
\author{
Christian Frei
}
