% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spgl.dt.R
\docType{methods}
\name{spgl.dt.lsFileStr.getDtDlDir}
\alias{spgl.dt.lsFileStr.getDtDlDir}
\alias{spgl.dt.lsFileStr.getDtDlDir,spgl.dt.lsFileStr-method}
\title{spgl.dt.lsFileStr.getDtDlDir}
\usage{
spgl.dt.lsFileStr.getDtDlDir(lsFileStr)

\S4method{spgl.dt.lsFileStr.getDtDlDir}{spgl.dt.lsFileStr}(lsFileStr)
}
\arguments{
\item{lsFileStr}{spgl.dt.lsFileStr}
}
\description{
create a method to get the value of the dtDlDir
}
\details{
XXX
}
\examples{
lsFileStr <- new("spgl.dt.lsFileStr")
print(spgl.dt.lsFileStr.getDtDlDir(lsFileStr))
}
\author{
Christian Frei
}
