% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spgl.dt.R
\docType{methods}
\name{spgl.dt.lsFileStr.getDtHeadlineFile}
\alias{spgl.dt.lsFileStr.getDtHeadlineFile}
\alias{spgl.dt.lsFileStr.getDtHeadlineFile,spgl.dt.lsFileStr-method}
\title{spgl.dt.lsFileStr.getDtHeadlineFile}
\usage{
spgl.dt.lsFileStr.getDtHeadlineFile(lsFileStr)

\S4method{spgl.dt.lsFileStr.getDtHeadlineFile}{spgl.dt.lsFileStr}(lsFileStr)
}
\arguments{
\item{lsFileStr}{spgl.dt.lsFileStr}
}
\value{
spgl.dt.lsFileStr
}
\description{
XXX
}
\details{
XXX
}
\examples{
lsFileStr <- new("spgl.dt.lsFileStr")
print(spgl.dt.lsFileStr.getDtHeadlineFile(lsFileStr))
}
\author{
Christian Frei
}
