% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spgl.dt.R
\docType{class}
\name{spgl.dt.lsFileStr}
\alias{spgl.dt.lsFileStr}
\alias{initialize,spgl.dt.lsFileStr-method}
\title{Class spgl.dt.lsFileStr}
\usage{
\S4method{initialize}{spgl.dt.lsFileStr}(.Object)
}
\arguments{
\item{.Object}{.Object}
}
\description{
XXX

Constructor of spgl.dt.lsFileStr Class
}
\details{
XXX

XXX
}
\section{Slots}{

\describe{
\item{\code{dtDlDir}}{dtDlDir}

\item{\code{dtHeadlineFile}}{dtHeadlineFile}

\item{\code{dtHeadline}}{dtHeadline}
}}

\seealso{
spgl.dt.dtDlDir, spgl.dt.dtHeadlineFile, spgl.dt.dtHeadline
}
\author{
Christian Frei

Christian Frei
}
