% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spgl.dt.R
\docType{methods}
\name{spgl.dt.lsFileStr.setDtDlDir}
\alias{spgl.dt.lsFileStr.setDtDlDir}
\alias{spgl.dt.lsFileStr.setDtDlDir,spgl.dt.lsFileStr,data.table-method}
\title{spgl.dt.lsFileStr.setDtDlDir}
\usage{
spgl.dt.lsFileStr.setDtDlDir(lsFileStr, dtDlDir)


  \S4method{spgl.dt.lsFileStr.setDtDlDir}{spgl.dt.lsFileStr,data.table}(lsFileStr,
  dtDlDir)
}
\arguments{
\item{lsFileStr}{spgl.dt.lsFileStr.setDtDlDir}

\item{dtDlDir}{data.table}
}
\value{
spgl.dt.lsFileStr spgl.dt.lsFileStr
}
\description{
create a method to get the value of the dtDlDir
}
\details{
XXX
}
\examples{
lsFileStr <- new("spgl.dt.lsFileStr")
dtDlDir <- spgl.dt.dtDlDir(id = as.integer(1),
                           forDate = as_date(today()),
                           dlDir = "a",
                           dlDir.fllPath = "b")
lsFileStr <- spgl.dt.lsFileStr.setDtDlDir(lsFileStr, dtDlDir)
print(spgl.dt.lsFileStr.getDtDlDir(lsFileStr))
}
\author{
Christian Frei
}
