% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spgl.dt.R
\docType{methods}
\name{spgl.dt.lsFileStr.setDtHeadline}
\alias{spgl.dt.lsFileStr.setDtHeadline}
\alias{spgl.dt.lsFileStr.setDtHeadline,spgl.dt.lsFileStr,data.table-method}
\title{spgl.dt.lsFileStr.setDtHeadline}
\usage{
spgl.dt.lsFileStr.setDtHeadline(lsFileStr, dtHeadline)


  \S4method{spgl.dt.lsFileStr.setDtHeadline}{spgl.dt.lsFileStr,data.table}(lsFileStr,
  dtHeadline)
}
\arguments{
\item{lsFileStr}{spgl.dt.lsFileStr}

\item{dtHeadline}{data.table}
}
\value{
spgl.dt.lsFileStr
}
\description{
XXX
}
\details{
XXX
}
\examples{
lsFileStr <- new("spgl.dt.lsFileStr")
dtHeadline <- spgl.dt.dtHeadline(id = as.integer(1),
                                 idHeadline = as.integer(2),
                                 idHeadlineFile = as.integer(3),
                                 forDate = as_date(today()),
                                 forTime = as_datetime(now()),
                                 isBentoLink = FALSE,
                                 isSpiegelPlusLink = FALSE,
                                 intro = "a",
                                 title = "b",
                                 section = "c",
                                 link = "d")
lsFileStr <- spgl.dt.lsFileStr.setDtHeadline(lsFileStr, dtHeadline)
spgl.dt.lsFileStr.getDtHeadline(lsFileStr)
}
\author{
Christian Frei
}
