% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spgl.dt.R
\docType{methods}
\name{spgl.dt.lsFileStr.setDtHeadlineFile}
\alias{spgl.dt.lsFileStr.setDtHeadlineFile}
\alias{spgl.dt.lsFileStr.setDtHeadlineFile,spgl.dt.lsFileStr-method}
\title{spgl.dt.lsFileStr.setDtHeadlineFile}
\usage{
spgl.dt.lsFileStr.setDtHeadlineFile(lsFileStr, dtHeadlineFile)

\S4method{spgl.dt.lsFileStr.setDtHeadlineFile}{spgl.dt.lsFileStr}(lsFileStr,
  dtHeadlineFile)
}
\arguments{
\item{lsFileStr}{spgl.dt.lsFileStr}

\item{dtHeadlineFile}{dtHeadlineFile}
}
\value{
spgl.dt.lsFileStr spgl.dt.lsFileStr
}
\description{
XXX
}
\details{
XXX
}
\examples{
lsFileStr <- new("spgl.dt.lsFileStr")
dtHeadlineFile <- spgl.dt.dtHeadlineFile(id                   = integer(1),
                                         idDlDir              = integer(1),
                                         forDate              = lubridate::as_date(today()),
                                         inDate               = lubridate::as_date(today()),
                                         inTime               = lubridate::as_datetime(now()),
                                         dlDir                = "a",
                                         dlDir.fllPath        = "b",
                                         rawDir               = "c",
                                         rawDir.fllPath       = "d",
                                         fileHTML             = "e",
                                         fileHTML.fllPath     = "f",
                                         fileCSV              = "g",
                                         fileCSV.fllPath      = "h")
lsFileStr <- spgl.dt.lsFileStr.setDtHeadlineFile(lsFileStr, dtHeadlineFile)
print(spgl.dt.lsFileStr.getDtHeadlineFile(lsFileStr))
}
\author{
Christian Frei
}
